수정 및 lv.6 구현 중

+++ src/main/java/com/plannerv2/todo/service/TodoService.java
@@ -22,24 +22,13 @@ public class TodoService {
 
     // 일정 생성
     @Transactional
-    public CreateTodoResponse crateTodo(CreateTodoRequest request) {
-        User user = userRepository.findById(request.getUserId()).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "없는 유저입니다."));
-        Todo todo = new Todo(
-                user,
-                request.getTitle(),
-                request.getContent()
-        );
+    public CreateTodoResponse crateTodo(Long userId, TodoRequest request) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "유저를 찾을 수 없습니다."));
+        Todo todo = new Todo(user, request.getTitle(), request.getContent());
         Todo crateTodo = todoRepository.save(todo);
 
-        return new CreateTodoResponse(
-                crateTodo.getId(),
-                crateTodo.getUser().getId(),
-                crateTodo.getTitle(),
-                crateTodo.getContent(),
-                crateTodo.getCreatedAt(),
-                crateTodo.getModifiedAt()
-        );
+        return new CreateTodoResponse(crateTodo.getId(), crateTodo.getUser().getId(), crateTodo.getTitle(), crateTodo.getContent(), crateTodo.getCreatedAt(), crateTodo.getModifiedAt());
     }
 
     // 일정 조회(단 건)
@@ -48,14 +37,7 @@ public class TodoService {
         Todo todo = todoRepository.findById(id).orElseThrow(
                 () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "해당 일정이 존재하지 않습니다.")
         );
-        return new GetTodoResponse(
-                todo.getId(),
-                todo.getUser().getId(),
-                todo.getTitle(),
-                todo.getContent(),
-                todo.getCreatedAt(),
-                todo.getModifiedAt()
-        );
+        return new GetTodoResponse(todo.getId(), todo.getUser().getId(), todo.getTitle(), todo.getContent(), todo.getCreatedAt(), todo.getModifiedAt());
     }
 
     // 전체 일정 조회
@@ -65,14 +47,7 @@ public class TodoService {
 
         List<GetTodoResponse> dtos = new ArrayList<>();
         for (Todo todo : todos) {
-            GetTodoResponse dto = new GetTodoResponse(
-                    todo.getId(),
-                    todo.getUser().getId(),
-                    todo.getTitle(),
-                    todo.getContent(),
-                    todo.getCreatedAt(),
-                    todo.getModifiedAt()
-            );
+            GetTodoResponse dto = new GetTodoResponse(todo.getId(), todo.getUser().getId(), todo.getTitle(), todo.getContent(), todo.getCreatedAt(), todo.getModifiedAt());
             dtos.add(dto);
         }
         return dtos;
@@ -80,25 +55,28 @@ public class TodoService {
 
     // 일정 수정
     @Transactional
-    public UpdateTodoResponse updateTodo(Long id, UpdateTodoRequest request) {
+    public UpdateTodoResponse updateTodo(Long id, Long userId, TodoRequest request) {
         Todo todo = todoRepository.findById(id).orElseThrow(
                 () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "해당 일정이 존재하지 않습니다.")
         );
 
+        if (!todo.getUser().getId().equals(userId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "본인의 일정만 수정할 수 있습니다.");
+        }
+
         todo.updateTodo(request.getTitle(), request.getContent());
-        return new UpdateTodoResponse(
-                todo.getTitle(),
-                todo.getContent(),
-                todo.getModifiedAt()
+        return new UpdateTodoResponse(todo.getTitle(), todo.getContent(), todo.getModifiedAt()
         );
     }
 
     // 일정 삭제
     @Transactional
-    public void deleteTodo(Long id) {
+    public void deleteTodo(Long id, Long userId) {
         Todo todo = todoRepository.findById(id).orElseThrow(
-                () ->  new ResponseStatusException(HttpStatus.NOT_FOUND, "해당 일정이 존재하지 않습니다.")
-        );
+                () ->  new ResponseStatusException(HttpStatus.NOT_FOUND, "해당 일정이 존재하지 않습니다."));
+        if (!todo.getUser().getId().equals(userId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "본인의 일정만 수정할 수 있습니다.");
+        }
         todoRepository.deleteById(id);
     }
 }
